Update Clothing Assignments of Suit/Skirt (#857)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Changes some outfits, notably the Operative Jumpskirt, Kimono,
Dark/Light Monastic Robes, and the Gladiator Uniform to be assigned as
skirts, rather than suits. This allows harpies to wear them. In
addition, the UnsensoredClothingSkirtBase base now possesses the
'ClothMade' tag, allowing moths to eat outfits with this base.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Fixes a few typos in outfit descriptions
- fix: The Kimono, Dark/Light Monastic Robes, Gladiator Uniform, and
Operative Jumpskirt should be properly considered jumpskirts, and are by
extension wearable for harpies.
- fix: Moths should be able to eat outfits with the
UnsensoredClothingSkirtBase parent.