Service & Instrument Loadout Groups (#906)
# Description

This PR creates Loadout Groups specific to Service jobs, and provides a
discount on all service loadout items in return for having the group
limit behavior. Also adds significant discounts to all instruments for
musicians in exchange for only being able to pick up to 3.

---

# Changelog

:cl: Mocho
- add: Service Job related items are now part of a Service loadout
group, and have all been significantly discounted.
- tweak: Musicians now have significant discounts on all instruments in
loadouts, but can only take a maximum of 3.

Signed-off-by: gluesniffler <159397573+gluesniffler@users.noreply.github.com>
# Conflicts:
#	Resources/Prototypes/Loadouts/Jobs/service.yml