Reverse Engineering Update Pr Port/ Board Cost Balance  (#910)
Ports the re update from delta and adds the Rcd and its ammo to the re
machine along with jaws of life for more potential usage. Also moves the
Re components to the Bottom of each yml item for items with the Re
component placed in the middle likely from time and neglect, so it's
consistent and less messy and better to understand. Also makes the cost
of the naynoboards consistent with the other boards


Expands the re machines catalog and adds it to alternate research tech
as well as making it craft able via plasma glass, makes all Nanyo-boards
the same cost as the ones from the hyper lathe updates such
I think this would help those who wish to work on the re machine helps
with future updates I need to flesh out too such as the other syndicate
suits and items, adds a little more use and makes it consistent with
another tech in nt's catalog. The Remechine has been neglected for a
long time, and it seems that super parts are in the work here so ill
likely make the suits and wait for the updates rather than making a
super re-machine like planed.
--
Media 

![Screenshot 2024-09-11
114101](https://github.com/user-attachments/assets/b337ee2b-e851-40c4-9377-8ff17c932908)
![Screenshot 2024-09-11
114140](https://github.com/user-attachments/assets/3f75f8ca-ca04-4670-b254-cdd6b6f703be)
![Screenshot 2024-09-11
113137](https://github.com/user-attachments/assets/00b31c57-6d8d-46ba-8887-0e6e2a55269d)


---

# Changelog


:cl:
- add: added rcds to the list of possible re and ported over the re
update from delta
- tweak: tweaked the costs of boards re-machine boards to be more
consistent with their post hyper lathe update counter parts

---------

Signed-off-by: Squishy77 <165165480+Squishy77@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>