Bartender Loadout Groups (Also Item Group Localizations) (#915)
# Description

This PR adds Bartender-Specific loadout item groups. Very simple, it
removes some of the items from the Bartender's Locker, and replaces them
with giving the bartender a choice of which "Bar Gun" he wishes to start
with. Currently just the classic beanbag double barrel, but now the
bartender can elect to spend some of his loadout points to instead start
with a Mosin(Preloaded with Rubber bullets)

Also while I was at it, I added missing localizations.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/ca51fc69-966e-429e-b035-9c6d4f6d2dc9)

</p>
</details>

# Changelog

:cl:
- add: Added missing Localizations for Departments, Jobs, and Loadout
Item Groups.
- add: Bartender-Specific loadouts! Bartenders can now choose to spawn
with either their classic shotgun, or a Mosin Nagant that comes
preloaded with rubber bullets.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>