Corpsman Glasses (#1828)
* Corpsman Glasses

Added Corpsman Glasses to the game, added it to Corpsmans lockers, and added a crafting recipe for them.

* Corpsman Glasses

Added Corpsman Glasses to the game, added it to Corpsmans lockers, and added a crafting recipe for them.

* Corpsman Glasses

Added Corpsman Glasses to the game, added it to Corpsmans lockers, and added a crafting recipe for them.

* Corpsman Glasses

Added Corpsman Glasses to the game, added it to Corpsmans lockers, and added a crafting recipe for them.

(cherry picked from commit d0ee16144139954708aa42bb7ae56c5fe456a9e8)