IPC Refactor (#771)
# Description

IPCs were ported from a codebase that didn't necessarily follow all of
our repo's coding standards. And while I had done my part to cleanup as
much of the system as was practical within the bounds of a Maintainer
Review, there were a lot of things that I felt were inappropriate to
leave to review, and wished to go over with a fine lense. Thus, here is
my Refactor of IPC code.

Do not merge this without first testing that nothing was broken. Because
I haven't tested it myself yet.