Psionic "Heal Other" Powers (#942)
# Description

This PR introduces two new Psionic Powers, Healing Word, and Breath of
Life, both utilizing a new PsionicHealOtherSystem, which operates on
datafield event arguments rather than a "hardcoded" component. Thus, any
number of powers can be created which share this system.

Healing Word is a power that features a short cast time, and heals a
small amount of each damage type to a target(while reducing the target's
rot timer slightly). It has a relatively short cooldown, and a low
glimmer cost.

Breath of Life by contrast, is an extremely rare power with a longer
cast time, healing a much larger amount of each damage type to a target,
reduces rot significantly, and attempts to revive the target. It has a 2
minute cooldown, and a high glimmer cost.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/ba01ccce-639f-4b03-84bb-55f96b5aeda3)

</p>
</details>

# Changelog

:cl:
- add: Healing Word has been added as a new Psionic Power. When cast on
another person, it heals a small amount of every damage type(scaling
with Casting Stats), while also reducing rot timers. Healing Word has a
very short cooldown, and a fairly low Glimmer cost.
- add: Breath of Life has been added as a new extremely rare Psionic
Power. When cast on another person, it heals a large amount of
damage(scaling with Casting Stats), while also substantially reducing
rot timers. Additionally, it will revive the target if it is possible to
do so. Breath of Life has an incredibly long cooldown, a long
interuptable cast time, and an extraordinarily high glimmer cost(A
typical Psion will spike glimmer by more than 50 points when casting
it).
- add: The Chaplain now starts with the Healing Word power.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>