More Objects Hit By Projectiles (#943)
# Description

This PR is a follow up to
https://github.com/Simple-Station/Einstein-Engines/pull/939
By increasing the number of items that are struck by bullets, or can be
targeted by bullets. To add to the "Cinematic" experience of a gunfight,
Bottles, Cups, and Plates of all kinds are automatically struck by
bullets when fired over, and *probably* destroyed given that they only
have 5 hit points. This also serves to aid in limiting the amount of
things that can be hidden underneath when crawling under a table, since
the station's bar counter will likely explode into a ton of glass and
spilled booze when a gunfight gets near it.

I also added Chairs to the "RequireProjectileTarget" feature, so that if
you click on a chair, you'll shoot the chair and (probably) destroy it.

# Changelog

:cl:
- add: Bottles, Drink Glasses, Plates, and all liquid containers are now
struck by bullets(and most likely destroyed in the process). We hope
that this will offer both a small tactical advantage/disadvantage, as
well as contribute to making gunfights around the bar more "Cinematic".
- add: Chairs are now hit by projectiles if a shooter clicks on them, in
addition to Tables.