Fix HealingWord & Revivify Bugs (#949)
# Description

Playtesting uncovered a few bugs with the HealOtherSystem, these are
luckily pretty simple to fix. I also made a few of the functions in this
system nullable as a performance optimization.

# Changelog

:cl:
- fix: Fixed a bug where Breath of Life and Healing Word would set
living people to crit, which would knock them down.
- fix: Fixed a bug where Breath of Life could revive people who were
rotten(It will still heal them and reduce the rot timer as intended, but
will not revive unless said rot timer reduction brings them below the 10
minute threshold).
- add: The Do-After bar for Breath of Life and Healing Word is now
hidden if glimmer is low enough. The threshold for which scales with
your Dampening stat. More Dampening = stealthier casting.