More Quick Construction Recipes And Upgrades (#945)
# Description

This PR adds a few new recipes to Quick Construction (accessed by
pressing Z on commonly used construction materials/items, such as rods,
glass, steel, plasma, etc etc). Along with simple nesting for grouping
items up under categories, akin to how the RCD works. Which allows
engineers to massively speed up pipe laying work, disposal rebuilding,
setting up the supermatter, etc etc.

First time doing UI code, apologies for the shitcode in advance :godo:

---

<details><summary><h1>Media</h1></summary>
<p>

https://github.com/user-attachments/assets/9d540d7f-7e4f-4a43-874d-5ea069011a37

</p>
</details>

---

# Considerations

Is this powergamey? Probably. Is it fucking awesome? Hell yeah. Some
servers might question the fact that everyone can build stuff this fast
now that they dont have to deal too much with the construction menu for
the most common things... And might want to lock it behind traits or
given as a job freebie, like CPR for doctors.

Other servers might appreciate it if they are looking for more action to
happen over the round, and giving everyone the ability to deal with
breaches and repairs quicker is probably a step into that direction.
I'll be observing for feedback and adjust accordingly.

It also shouldn't invalidate any eventual RPDs, or other devices of the
kind since this is just a shortcut for construction, you still require
materials and a doafter for most of them.

---

# Changelog

:cl: Mocho
- add: Added a lot of recipes to the quick construction menus. Give it a
shot by pressing Z with different construction materials in your hand!

---------

Signed-off-by: gluesniffler <159397573+gluesniffler@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
# Conflicts:
#	Content.Client/ShortConstruction/UI/ShortConstructionMenuBUI.cs