Backpack Loadout Category (#957)
# Description

https://github.com/Simple-Station/Einstein-Engines/pull/794 apparently
silently deprecated the Backpack Preferences button. Selecting a
backpack preference no longer actually does anything, and instead your
backpack comes from your Loadout. This would have been fine if we were
warned about this, but because nobody merged upstream for 2 weeks after
the release of PR794, this issue went unreported.

Well now the solution is going to require adding every backpack in the
entire game to a new Backpacks tab. I've started this PR, but because
it's really late and I have Calculus classes in the morning, I won't be
able to finish it tonight. So I'm enlisting the help of a few other
people I know to PR to this PR to help get the backpacks in.

After we have backpacks in the loadout tab, and after Wizmerge, I'll
probably do a separate PR to actually remove the now useless Backpack
selection button.

This is probably for the better long term. Since this technically means
Backpack slots are no longer hardcoded to "Backpack/Duffelbag/Satchel",
and now the back slot is freeeeee to be used for anything.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/7081660f-68f0-471e-99eb-67f026b7e122)

</p>
</details>

# Changelog

:cl:
- add: Backpacks, Duffelbags, and Satchels have been added to loadouts.
- remove: The button for "Backpack Preference" has been removed.
Backpacks selection is now done via Loadouts.
- remove: The button for "Suit/Skirt Preference" has also been removed.
Suit selection is done via Loadouts.