Harpy Language! (#968)
# Description

This PR adds a new Language specific to the Harpy species, which is
composed of syllable charts taken from Polish, Serbian, Greek, and
Turkish. It uses Cambria as its font, which is modestly different from
the standard font, but still reasonably readable. This also adds a
Language menu trait that allows people to buy the language with a trait
point.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/929a5110-2339-4410-a97d-9f1bcf537e7b)

</p>
</details>

# Changelog

:cl:
- add: Harpies now have their own species-specific language, called
Valyrian Standard
- add: Valyrian Standard can now be bought as an extra language in the
Traits menu.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>