Anomalist Power System (#991)
# Description

This PR extends the system originally created for Shadeskip, into a
system that allows any (currently existing) anomaly effect to be
replicated as an Instant Action psionic power. This will be needed for
Wizard/Rogue Psion antagonists. I'll be making a metric shitton of new
powers using this in separate PRs.

This is extremely blatantly re-using code as-is from the various Anomaly
Systems, because it unfortunately turns out that it's simply not
possible to build a constructor to interface between Psionics and
Anomalies.

# Changelog
No changelog because this isn't player facing.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>