MoodSystem Crit Threshold CVar (#1010)
# Description

This PR adds a CVar that allows server hosts (such as N14 who need it)
to optionally disable the Crit Threshold modification part of the
MoodSystem. This is useful if a server has some other system that
frequently messes with Thresholds, and that a server host wishes their
system to not have to fight with Mood for it.

# Changelog

No changelog because this isn't playerfacing.