Replace Shittle With Pathfinder (#1027)
# Description

Part 3 of my evil plan to make salvage not suck: Give them a fancy
goody, but with a twist. This PR adds the Pathfinder Expedition Vessel
from Frontier, which replaces the Shittle as the default salvage shuttle
to spawn on each map. I have my own funky reasoning for this. The
primary one being, "The Pathfinder is powered by an AME. This means
salvage are REQUIRED to periodically return to the station to get fuel.
It is impossible for them to refuel themselves away from the station."

It also, notably, spawns with an Expedition Console. More on that in a
follow-up PR where I rework expeditions to work "More like Frontier".

# TODO

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/f516ffdc-585b-4e00-83b3-f1fd78b3ffd1)


![image](https://github.com/user-attachments/assets/8c6bffcb-f6a7-4963-93a8-8f6d90957eb8)

</p>
</details>

# Changelog

:cl:
- add: The Salvage Shittle has been replaced with the Pathfinder
Expedition Vessel from Frontier.