Bluespace And Normality Ores (#1042)
# Description

This PR adds new ores in the form of Bluespace and Normality ores, the
former being just as rare as artifact fragments, and the latter being
even rarer still. These ores can be mined by salvage, will generate on
asteroids or expeditions, and can be processed in an ore processor into
bluespace or normality crystals, which Epistemics can use to create
psionic related objects. Note that the two are currently
indistinguishable from the other until you actually mine them.

Yes this DOES infact make it easier to make glimmer drains and probers.
I need this as an excuse to make psionics even more dangerous.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/2c7b38a4-d700-43d6-ad47-a063839f0540)


![image](https://github.com/user-attachments/assets/5da73dff-acca-4395-855d-75a04f369ead)


![image](https://github.com/user-attachments/assets/269e231d-d702-4387-9c89-f8aec7854fb3)

</p>
</details>

# Changelog

:cl:
- add: Added Bluespace and Normality ore as very rare ore types. Salvage
can occasionally find these ores on Asteroids or on Expedition planets.
Bluespace and Normality Ore can be smelted in an Ore Processor into
Bluespace and Normality Crystals. Make sure to bring these to Epistemics
so that they can do their job.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>