Fix Melee Contests (#1069)
# Description

Default values for melee weapon contests were causing balance issues
with Mobs that interacted with it, such as Carps being able to defeat
the damage threshold of reinforced walls when the intention is that they
shouldn't be able to. Also the logic on ContestDisadvantages was
inverted, so this fixes that.

# Changelog

:cl:
- fix: Fixed issues with Contests System that allowed mobs such as Space
Carps to smash through barriers that were not intended to be breakable
by said mobs.

(cherry picked from commit 7f60931aa06a7bf490bcfb8865e12bbcd3b02265)