Fix Cryptobiolin (#1081)
# Description

Closes #1080
This PR fixes Cryptobiolin so that it correctly adds a temporary
"PsionicsDisabled" and "PsionicallyInsulated" status effect to the
consumer. This effect lasts for 15 minutes, and is intended to be a
highly effective means of providing protection from psionics, especially
in the case of a Code White call.

# Changelog

:cl:
- fix: Fixed Cryptobiolin not providing temporary Psionic Insulation. It
now provides protection from psionic abilities and events for 15
minutes.

(cherry picked from commit 528ef89420dbf3eb240d35f8686ff98ff7d26307)