Fix Steal Hos Weapon Objective (#1084)
# Description

Closes #1076

This fixes the "Steal The HoS's weapon" objective by first making it a
SINGLE objective that allows to stealing any of the HoS Weapons,
regardless of which one was selected. It also toggles on the
"VerifyMapExistence" setting for this objective, so that it will
correctly not generate if the station's Head of Security either isn't
present, or simply chose not to take an antique weapon to the station(I
don't know why they would do this, but it's an option).

# Changelog

:cl:
- fix: Fixed the "Steal The Head of Security's Weapon" so that it is now
a single objective that counts for any anique weapon chosen by the Head
of Security. The objective will also once again correctly not generate
if there are no valid items for it on the station.

(cherry picked from commit ba171b935fb0251a2464099544d59ce5473e6df5)