Port Sacrificing (#1110)
# Description

This PR rebases the Sacrificing mechanic from Nyanotrasen, while
updating it to newer standards for the Reworked Psionics System. This
does not rebase Golemancy, or Items Of Power, so the system for now
exists solely to spike glimmer DOWN in the event of a code white, and
merely produces brains to be placed inside cyborgs. Notably different
from DeltaV's implementation of it, this system does not reduce a random
amount of glimmer, and instead reduces glimmer by a base of 25,
multiplied by the victim's Psionic Amplification(which is itself
semi-random). Which means that certain individuals(Especially people
with lots of powers, and/or the Power Overwhelming trait) are inherently
more useful to sacrifice than other, weaker psionics.

When I bring back Items Of Power, the chances of obtaining an Artifact
are based on the victim's Dampening stat.

https://github.com/user-attachments/assets/23e52f20-ac3e-4bd9-8122-cadec96cbdf3

# Changelog

:cl:
- add: Sacrificing Psions has been added. Psions can be sacrificed by
Epistemics upon an altar in order to dramatically reduce glimmer. The
more powerful the Psion to be sacrificed, the more glimmer is reduced.

---------

Co-authored-by: deltanedas <39013340+deltanedas@users.noreply.github.com>
(cherry picked from commit c557d74ffe4d97af4148ae36b22c54080ccd558c)