Updade RadStation (#1129)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

this PR fix #1126

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![imagem_2024-10-22_112221053](https://github.com/user-attachments/assets/d9b21d1e-834b-4c8a-a27e-e3d6ea079567)

![imagem_2024-10-22_112257222](https://github.com/user-attachments/assets/58a99618-ed63-4193-9ee0-1d1d83cc9d5f)

![imagem_2024-10-22_112309324](https://github.com/user-attachments/assets/bf4d2e9b-6b9c-4de5-b5f4-d53428fd02fe)

![imagem_2024-10-22_112323017](https://github.com/user-attachments/assets/97efceb6-a4f5-435f-abb9-3115ce1e2870)

![imagem_2024-10-22_112333040](https://github.com/user-attachments/assets/bea13c3d-6278-4b53-ae11-070782d4c741)

![imagem_2024-10-22_112343862](https://github.com/user-attachments/assets/88ead88e-5db0-46eb-a4cb-480946ed96cc)

![imagem_2024-10-22_112417214](https://github.com/user-attachments/assets/cec33b4d-8bba-4872-8573-3d224d526b06)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: #1126

(cherry picked from commit e46ed866649ccefb078baab807de3d6de2b49dc5)