Updade Gax (#1163)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR updates the Gax map, completely modifying the AI ​​Core and
preparing the location for the future addition of AI to the game.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Pipes, scrubbers, air vent and air alarms
- [x] Layout
- [x] Energy sistem
- [x] New Telecomms
- [x] Minor things
- [x] Cameras
- [x] Map images

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

Old version:

![image](https://github.com/user-attachments/assets/3ca6fce1-7461-4c54-82c9-23192291afdb)

new:

![image](https://github.com/user-attachments/assets/f15901ea-bda1-4c3e-96f3-389dfd7673eb)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Fixed several issues with the Gax Station. It also now has a
newly redesigned AI Satellite.

---------

Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit d2bdbec619f21e0b78efb94b715ed8d04ec787ec)