Fixes Cocoons (#1169)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR fixes the cocoon muffle issue and also add a "uncocoon" feature
for cocooners to use.
This also add icons on cocoon and fixes a butcher variable.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/59c6dd7e-6b2b-45e4-bbab-e02db5d9b9ce)

![image](https://github.com/user-attachments/assets/574b9c7b-7e78-403c-82a4-cc0c3a7f2aaf)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: FoxxoTrystan
- tweak: Cocooners can now freely uncocoon any cocoons without the need
of a sharp object.
- fix: Cocoon making you stuck in muffle accent.

(cherry picked from commit 3bbcab8ba504896ce41c1bc236b6a31274504bc4)