Add New Sprites (#1174)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR adds new prites for lockers, wall lockers and PDA.

the new PDA sprites are to be added to the loadout for command jobs in
the future

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/ddc0afd0-65d4-473c-b8bc-d941d2409447)

![image](https://github.com/user-attachments/assets/1e5bc6fb-da03-41e0-94a4-f3cf982dac82)

![image](https://github.com/user-attachments/assets/34f883cb-0d75-4bcb-bcf4-4cf219cf22e9)

![image](https://github.com/user-attachments/assets/4610fd69-91a9-44d0-93d3-702445882674)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: lockers for Syndicate Agent (for mapping), Janitor and Bartender
have been added.
- add: a new wall variant has been added for the radiation closet
- add: Uranium Crate added

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit b553ce9ae9b95feecc8fb9a1ad9f70c2e3182793)