Traits V2 (#1182)
# Description

Traits Version 2 implements a much needed overhaul of the Trait System,
changing it so that it no longer uses hardcoded nullable lists of
various entries required for the trait functions, and instead accepts a
*single* Modular List of Functions, and then iterates through the list
like a set of instructions on how to modify the character. This PR also
features a few new Functions, such as AddImplant.

Anyway, rather than hardcoding a new codepath for every new function,
new functions can simply be made separate from the codepath entirely,
and will only be called if they are included in the trait itself.

# TODO

- [x] Remove ALL pre-existing trait functions and move them over to the
new system.
- [x] Evaluate which System Dependencies are essential for a majority of
functions, and make sure that TraitSystem is caching them on behalf of
the functions that need them for performance reasons. This system has a
fairly high startup cost for frametime to begin with.
- [x] Update all pre-existing traits such that they use the new System.

# BREAKING CHANGES

ALL DOWNSTREAM TRAITS WILL NEED TO BE MODIFIED TO USE THE NEW SYSTEM.

# Changelog

:cl:
- add: Traits can now add Implants directly.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: gluesniffler <159397573+gluesniffler@users.noreply.github.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>
(cherry picked from commit e7d40d121cc95b43538c3b6a55119e5778108198)