Give Penguins Cold Resistance (#1229)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR is intended to fix the issue of pengiuns dying to cold damage on
planetary maps like Glacier. It adds a
atmosTemperatureTransferEfficiency modifier to the pengiun mod YAML.

Reported on Grimbly, fix from
[Floof](https://github.com/Fansana/floofstation1/pull/302/files)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->
n/a

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

n/a

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: zelezniciar
- fix: Fixed pengiun cold resistance

(cherry picked from commit 71185a7e7ff12d91b30f1b319107786b4e77ac8d)